Extract .achx data model into MonoGame-free FlatRedBall2.Animation.Content
AnimationEditor.Core only ever needed the pure .achx save-format types
(AnimationFrameSave, AnimationChainSave, ShapesSave, ColorOperation, and the
XML load/save members of AnimationChainListSave), but pulled in the entire
FlatRedBall2 engine -- and its net10.0 MonoGame.Framework.DesktopGL
dependency -- because those types shared a file/assembly with the runtime
bridge (ToAnimationChainList/BuildList, which needs Texture2D/ContentLoader).

Split the pure data model into its own project and moved the runtime bridge
into AnimationChainListSaveExtensions (extension methods, same call-site
syntax) in the main engine assembly. AnimationEditor.Core now references only
the new Content project, dropping its MonoGame dependency entirely -- the
prerequisite for #535 (Animation Editor on Avalonia.Browser/WASM).

Pure structural move: 1093 FlatRedBall2.Tests + 1433 AnimationEditor.Core.Tests
stayed green throughout, identical counts before and after.
