+++ src/Animation.Content/AnimationChainSave.cs
