+++ src/Animation.Content/ColorOperation.cs
@@ -1,8 +1,8 @@
 namespace FlatRedBall2.Animation;
 
 /// <summary>
-/// How a frame's per-frame color (<see cref="AnimationFrame.Red"/>/<see cref="AnimationFrame.Green"/>/
-/// <see cref="AnimationFrame.Blue"/>) combines with the sprite's texture. A <c>null</c> operation
+/// How a frame's per-frame color (<c>AnimationFrame.Red</c>/<c>AnimationFrame.Green</c>/
+/// <c>AnimationFrame.Blue</c>) combines with the sprite's texture. A <c>null</c> operation
 /// means none. Like the channels themselves, this is authored in the Animation Editor and stored in
 /// the <c>.achx</c>, but whether a given runtime applies it is that runtime's choice — the <c>.achx</c>
 /// is a general-purpose format consumed by several renderers (Gum, MonoGame/FNA, FRB1, FRB2).
