+++ tools/AnimationEditorAvalonia/src/AnimationEditor.Views/Controls/DrawTimeOverlay.cs
