+++ tools/AnimationEditorAvalonia/src/AnimationEditor.Views/Controls/GuideCursorResolver.cs
