+++ tools/AnimationEditorAvalonia/src/AnimationEditor.Views/Controls/HandleCursorMapper.cs
