+++ tools/AnimationEditorAvalonia/src/AnimationEditor.Views/Controls/InspectableImage.cs
