+++ tools/AnimationEditorAvalonia/src/AnimationEditor.Views/Controls/PreviewControl.cs
