+++ tools/AnimationEditorAvalonia/src/AnimationEditor.Views/Controls/WireframeControl.cs
