+++ tools/AnimationEditorAvalonia/src/AnimationEditor.Views/Controls/ZoomPresetStepper.cs
