+++ tools/AnimationEditorAvalonia/src/AnimationEditor.Views/FrameColorFilter.cs
