+++ tools/AnimationEditorAvalonia/src/AnimationEditor.Views/FramePreviewOpacity.cs
