+++ tools/AnimationEditorAvalonia/src/AnimationEditor.Views/Services/ThumbnailService.cs
