+++ tools/AnimationEditorAvalonia/src/AnimationEditor.Views/Theming/CanvasPalette.cs
